fix(bundle): lazy-load JSON glTF parsing
Move JSON .gltf buffer parsing out of the always-loaded GLB path so binary glTF scenes do not pay for relative buffer resolution.
